Divide Parser into different classes for each command
The parsing of all commands is done by a single Parser class, with one
method for each class. This results in there being a large Parser class
that is difficult to understand and also does not follow separation of
concerns.

Let's split the Parser class into multiple smaller XYZCommandParser
classes, each for one command. This makes for more readable code and
shorter classes that follow separation of concerns.
