+++ src/test/java/seedu/manager/command/ViewCommandTest.java
@@ -3,7 +3,7 @@ package seedu.manager.command;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import seedu.manager.event.EventList;
-import seedu.manager.parser.Parser;
+import seedu.manager.parser.ParserManager;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -17,15 +17,15 @@ public class ViewCommandTest {
     public void setUp() {
         Command command;
 
-        command = new Parser().parseCommand("add -e Event 1 -t 2024-10-10 10:00 -v Venue A");
+        command = new ParserManager().parseCommand("add -e Event 1 -t 2024-10-10 10:00 -v Venue A");
         command.setData(eventList);
         command.execute();
 
-        command = new Parser().parseCommand("add -p Tom -e Event 1");
+        command = new ParserManager().parseCommand("add -p Tom -e Event 1");
         command.setData(eventList);
         command.execute();
 
-        viewCommand = new Parser().parseCommand("view -e Event 1");
+        viewCommand = new ParserManager().parseCommand("view -e Event 1");
         viewCommand.setData(eventList);
         viewCommand.execute();
     }
